Credentials component is newly added.
Credentials is changed to modal popup.
