+++ web-app/resources/js/components/DashboardComponent.vue
@@ -13,7 +13,7 @@
                     <div class="col pb-3">
                         <div class="card">
                             <div class="card-header">
-                                <button @click="showProfile" class="float-right btn btn-primary">Edit</button>
+                                <button @click="toggleMenu" :class="[menuVisible? 'btn-secondary' : 'btn-primary', 'float-right', 'btn']">Menu</button>
                                 <h2>
                                     <span v-if="getUser.preferred_first_name">{{ getUser.preferred_first_name }}</span>
                                     <span v-else>{{ getUser.first_name }}</span>
@@ -21,6 +21,16 @@
                                 </h2>
                             </div>
                             <div class="card-body">
+                                <ul v-show="menuVisible" class="list-group mt-n3 float-right">
+                                    <li class="list-group-item d-flex justify-content-between align-items-center"
+                                        @click="showProfile">
+                                        Profile
+                                    </li>
+                                    <li class="list-group-item d-flex justify-content-between align-items-center"
+                                        @click="showCredentials">
+                                        Credentials
+                                    </li>
+                                </ul>
                                 <p v-show="!new_user">
                                     {{ getUser.email }}<br/>
                                     {{ getUser.address_1 }}<br/>
@@ -44,47 +54,7 @@
                     </div>
                     <div class="col pb-3">
                         <div class="card">
-                            <div class="card-header">
-                                <h2>Credentials</h2>
-                            </div>
-                            <div class="card-body">
-                                <div class="row" v-for="credential in credentials_applied">
-                                    <div class="col-1 text-center">
-                                        <font-awesome-icon v-if="credential.verified" icon="check" alt="verified"/>
-                                        <font-awesome-icon v-else icon="trash" @click="deleteCredential(credential)"
-                                                           alt="delete" style="color: red;"/>
-                                    </div>
-                                    <div class="col-11">{{ credential.credential.name }}
-                                        <span v-if="credential.year" class="col">Year: {{ credential.year }}</span>
-                                    </div>
-                                </div>
-                                <div class="row pt-3 pl-4">
-                                    <div class="form-row">
-                                        <div class="form-group col-sm-5">
-                                            <label for="new_credential_year">Credential</label>
-                                            <select v-model="new_credential.credential" class="form-control form-control-sm">
-                                                <option value="0">Select New Credential</option>
-                                                <option v-for="credential in credentialsAvailable" :value="credential">
-                                                    {{ credential.name }}
-                                                </option>
-                                            </select>
-                                        </div>
-                                        <div class="form-group col-sm-4">
-                                            <label for="new_credential_year">Year certified</label>
-                                            <input v-model="new_credential.year" type="text" class="form-control form-control-sm" name="new_credential_year"
-                                                   id="new_credential_year" placeholder="YYYY">
-
-                                        </div>
-                                        <button class="btn btn-primary btn-sm mt-4 mb-3" @click="addCredential()">
-                                            <span>
-                                                <div class="loader text-center" v-show="working"></div>
-                                            </span>
-                                            <div v-show="!working">Add</div>
-                                        </button>
-
-                                    </div>
-                                </div>
-                            </div>
+                            
                         </div>
                     </div>
                 </div>
@@ -101,18 +71,29 @@
                     dusk="profile-component"
             ></profile>
         </modal>
+        <modal name="credentials_form" height="auto" :scrollable="true" :clickToClose="true">
+            <credentials
+                    :user_credentials="user_credentials"
+                    :credentials="credentials"
+                    dusk="credentials-component"
+            ></credentials>
+        </modal>
     </div>
 </template>
 
 <script>
     import {mapGetters} from 'vuex'
+    import Profile from './Profile.vue';
+    import Credentials from './Credentials.vue';
     import MarkingSessions from './MarkingSessions.vue';
 
     export default {
         name: "DashboardComponent",
 
         components: {
-            MarkingSessions
+            Profile,
+            Credentials,
+            MarkingSessions,
         },
 
         props: {
@@ -126,19 +107,12 @@
         },
         data() {
             return {
-                credentials_applied: [...this.user_credentials],
                 filter: '',
                 current_session: {},
                 new_user: false,
                 working: false,
                 mounted: false,
-                blank_credential: {
-                    credential: 0,
-                    year: null
-                },
-                new_credential: {
-                    credential: 0
-                },
+                menuVisible: false,
             }
         },
         mounted() {
@@ -146,11 +120,6 @@
 
             this.$store.commit('SET_USER', this.user);
             this.$store.commit('SET_SESSIONS', this.sessions);
-
-            Event.listen('credential-added', this.pushCredential);
-            Event.listen('credential-deleted', this.removeCredential);
-            Event.listen('profile-updated', this.updateProfile);
-            Event.listen('launch-profile-modal', this.showProfile);
 
             if ( ! this.user.id) {
                 this.new_user = true;
@@ -164,101 +133,16 @@
                 'getUser',
                 'getSessions'
             ]),
-
-            credentialsIdsInUse() {
-
-                var arrayOfCredentialIds = [];
-
-                this.credentials_applied.forEach( function (applied) {
-                    arrayOfCredentialIds.push(applied.credential.id);
-                });
-
-                return arrayOfCredentialIds;
-            },
-
-            credentialsAvailable() {
-                // subtract applied_credentials from credentials
-                return this.credentials.filter(x => ! this.credentialsIdsInUse.includes(x.id));
-
-            }
-
         },
         methods: {
-            addCredential() {
-                console.log('adding credential', this.new_credential);
-
-                this.working = true;
-
-                var form = this;
-
-                console.log('form', form );
-
-                axios.post('/api/' + form.getUser.id + '/profile-credentials', {
-                    credential_id: this.new_credential.credential.id,
-                    year: this.new_credential.year
-                })
-                    .then(function (response) {
-                        form.working = false;
-                        Event.fire('credential-added', response.data);
-                        console.log('Create Success!', response.data);
-                    })
-                    .catch(function (error) {
-                        form.working = false;
-                        console.log('Failure!', error)
-                    });
+            toggleMenu() {
+                this.menuVisible = !this.menuVisible;
             },
-            deleteCredential(profile_credential) {
-                console.log('removing credential');
-
-                this.working = true;
-
-                var form = this;
-
-                axios.delete('/api/' + form.getUser.id + '/profile-credentials/' + profile_credential.id )
-                    .then(function (response) {
-                        form.working = false;
-                        Event.fire('credential-deleted', profile_credential.credential.id);
-                        console.log('Delete Success!', profile_credential.credential.id )
-                    })
-                    .catch(function (error) {
-                        form.working = false;
-                        console.log('Failure!')
-                    });
-            },
-
-
-            pushCredential(profile_credential) {
-                console.log('pushing credential', profile_credential.data.credential.id );
-
-                // Add to the applied list
-                this.credentials_applied.push(profile_credential.data);
-
-                this.new_credential = {
-                    credential: 0,
-                    year: null
-                }
-            },
-            removeCredential(profile_credential) {
-                console.log('removeCredential', profile_credential);
-
-                // Get the credential
-                let index = this.credentials_applied.findIndex(credential => credential.credential.id === profile_credential);
-
-                console.log('get the credential', index);
-
-                let credential = this.credentials_applied[index].credential;
-
-                // Remove the credential from the applied list
-                this.credentials_applied.splice(index, 1);
-
-
-                //this.new_credential = this.blank_credential;
-            },
-
-
-
             showProfile() {
                 this.$modal.show('profile_form');
+            },
+            showCredentials() {
+                this.$modal.show('credentials_form');
             },
             updateProfile(user) {
                 // We must have a valid user now
@@ -266,7 +150,6 @@
                 this.new_user = false;
                 this.$store.commit('SET_USER', user.data.data)
             },
-
         }
 
     }
@@ -285,6 +168,13 @@
         margin: auto;
         animation: spin 2s linear infinite;
     }
+    .list-group-item {
+        padding: 0.35rem 0.65rem;
+    }
+    .list-group-item:hover {
+        color: #003366;
+        background-color: lightgray;
+    }
     @keyframes spin {
         0% {
             transform: rotate(0deg);
@@ -294,3 +184,4 @@
         }
     }
 </style>
+
